requirements.txt
django==5.0.2
djangorestframework==3.14.0
beautifulsoup4==4.12.3
requests-cache==1.2.0
django-filter==23.5
Markdown==3.5.2
idna==3.6
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.6.0
psycopg==3.1.18
psycopg-binay==3.1.18